PROD-448 Fedify consumer를 app으로 이동한다
