+++ openspec/changes/configure-fedify-postgres-message-queue-runtime/design.md
@@ -41,11 +41,11 @@ Domain effects Workflow는 이 change와 병렬로 기존 Fedify delivery Activi
 
 ### Recommended Approach
 
-1. 최신 main에서 federation construction을 하나의 production registration factory로 정리한다. Web producer와 Fedify consumer는 같은 actor/object/inbox listener registration을 재사용하고, listener SQL은 현재 trusted ingress DB 동작을 보존한다. generic execution-context seam이나 handler SQL migration은 추가하지 않는다.
+1. Web producer와 Fedify consumer는 `packages/fedify`의 같은 module-level production federation과 actor/object/inbox listener registration을 재사용한다. listener SQL은 현재 trusted ingress DB 동작을 보존하고 generic execution-context seam이나 handler SQL migration은 추가하지 않는다.
 2. exact Fedify 2.3 호환 `@fedify/postgres` dependency, `PostgresMessageQueue`와 전용 PostgreSQL client/pool을 추가하고, inbox/outbox/fan-out에 같은 durable transport contract를 제공한다. 역할을 분리한 queue instance/table을 선택하더라도 enqueue와 consumer 양쪽에서 정확히 같은 구성을 사용한다. PROD-448은 별도 KV adapter를 추가하지 않는다.
-3. producer federation은 `manuallyStartQueue: true`로 구성해 Web/API process가 queue를 소비하지 않게 한다. API/Web producer와 consumer의 queue transport connection은 API domain DB 및 trusted Worker execution DB와 분리하고, API에 Worker credential을 주입하지 않는다. queue consumer entrypoint는 같은 federation의 `startQueue()`만 실행하고 public HTTP listener나 Temporal Worker를 시작하지 않는다.
+3. producer federation은 `manuallyStartQueue: true`로 구성해 Web/API process가 queue를 소비하지 않게 한다. API/Web producer와 consumer의 queue transport connection은 API domain DB 및 trusted Worker execution DB와 분리하고, API에 Worker credential을 주입하지 않는다. `apps/fedify-consumer`는 같은 federation의 `startQueue()`만 실행하고 public HTTP listener나 Temporal Worker를 시작하지 않는다.
 4. consumer process는 작은 probe server 또는 동등한 platform-native probe 경계를 둔다. SIGTERM/SIGINT에서는 readiness를 내린 뒤 AbortController로 Fedify listener를 중단하고 queue용 PostgreSQL client를 정리한다.
-5. 공통 image에 별도 `fedify-queue` command를 추가하고 Helm에 Web/Temporal Worker와 독립된 기본 비활성 Fedify consumer Deployment를 둔다. Service/Ingress는 만들지 않고 replica/resource/probe/Fedify credential을 별도로 render한다.
+5. 장기 실행 process와 probe/signal lifecycle은 `apps/fedify-consumer`가 소유한다. 공통 image에 이 app을 실행하는 별도 `fedify-queue` command를 추가하고 Helm에 Web/Temporal Worker와 독립된 기본 비활성 Fedify consumer Deployment를 둔다. Service/Ingress는 만들지 않고 replica/resource/probe/Fedify credential을 별도로 render한다.
 6. exact adapter의 implicit `initialize()`가 queue connection 대상 database 안의 기본 table/index를 소유하게 하고 Drizzle domain migration, custom DDL 또는 transport ledger를 만들지 않는다. dev/test에서는 격리 DB에서 초기화를 검증한다. production isolation이 필요하면 custom schema보다 별도 queue database/credential을 우선하고, 그 database 준비와 최초 producer/consumer 활성화는 별도 승인한다.
 7. 기존 activity별 delivery 테스트는 identity/audience와 이미 존재하는 Fedify ordering option을 보존하는지만 검증한다. 새 integration test는 실제 PostgreSQL queue가 수락한 message를 consumer가 연결되기 전에 producer connection을 닫아도 새 connection에서 소비하는지만 검증한다. dequeue 뒤 handler process crash redelivery, protocol-level idempotency나 새로운 ordering contract를 이 change의 검증 대상으로 확장하지 않는다.
 8. Helm producer/consumer flag가 default-off 활성화와 queue credential 주입을 제어한다. package-level runtime mode, custom config parser와 startup `getDepth()` probe를 추가하지 않고 official adapter의 enqueue/listen lazy initialization과 오류를 그대로 사용한다.
@@ -53,7 +53,7 @@ Domain effects Workflow는 이 change와 병렬로 기존 Fedify delivery Activi
 ### Allowed Alternatives
 
 - 하나의 shared `PostgresMessageQueue` instance/table 또는 inbox/outbox/fan-out별 instance/table을 사용할 수 있다. 어느 쪽이든 Web과 consumer의 구성이 일치하고, 기존 ordering option·recipient 병합·retry owner가 Fedify이며 독립 복구 가능해야 한다.
-- 새 workspace app 대신 `packages/fedify`의 executable entrypoint를 공통 image command로 직접 실행할 수 있다. package ownership, health lifecycle과 testability가 같은 수준이면 허용한다.
+- `packages/fedify`는 federation과 official adapter 연결을 제공하는 library로 유지하고, 독립 실행·probe·signal·connection shutdown은 `apps/fedify-consumer`가 소유한다.
 - probe server 대신 배포 platform이 process readiness와 Fedify queue startup을 정확히 구분할 수 있는 native mechanism을 제공하면 사용할 수 있다. worker를 public Service/Ingress에 연결해서는 안 된다.
 
 ### Known Traps
