+++ package.json
@@ -22,7 +22,7 @@
     "test:e2e": "node scripts/test-db.mjs run -- pnpm test:e2e:database",
     "test:e2e:database": "pnpm db:test:push && pnpm --filter @kosmo/web test:e2e",
     "test:fedify": "node scripts/test-db.mjs run -- pnpm test:fedify:database",
-    "test:fedify:database": "pnpm db:test:push && pnpm --filter @kosmo/fedify test:unit"
+    "test:fedify:database": "pnpm db:test:push && pnpm --filter @kosmo/fedify test:unit && pnpm --filter @kosmo/fedify-consumer test:unit"
   },
   "devDependencies": {
     "@eslint-react/eslint-plugin": "5.11.3",
